+++ source/timer.h
@@ -3,11 +3,11 @@
 #include <stdint.h>
 
 typedef struct{
-   uint16_t days;
-   uint8_t hours;
-   uint8_t minutes;
-   uint8_t seconds;
-   uint8_t frames;
+   uint32_t days;
+   uint32_t hours;
+   uint32_t minutes;
+   uint32_t seconds;
+   uint32_t frames;
 }timer;
 
 void timer_tick(timer& timer_to_tick);
